+++ packages/netlify-cms-editor-component-image/src/index.js
@@ -17,6 +17,9 @@ const image = {
       label: 'Image',
       name: 'image',
       widget: 'image',
+      media_library: {
+        allow_multiple: false,
+      },
     },
     {
       label: 'Alt Text',